fix(doc-150 item 6): close 4 storage-quota defects
Four follow-up fixes on the doc-150 item-6 storage-quota changeset:

1. /v1/agent-sessions bypassed the per-account storage-quota gate. The
   gate ran only in routes/sessions.ts resolveProfileBinding, so a
   profile-backed POST /v1/agent-sessions create (which dispatches a
   profile → mints the R2 sealed-blob save-back URL → grows size_bytes)
   skipped it. An over-cap account could keep launching via this path.
   Now the create resolves the OWNER's tier (self → caller tier;
   team-scoped → looked up via authRepo, threaded into the route deps)
   and calls assertWithinStorageQuotaForLaunch after the ownership check,
   mirroring routes/sessions.ts exactly. Enterprise stays soft-only.
   Adds integration tests: over-cap → 409 storage_quota_exceeded; under
   cap → 201; enterprise over-cap → 201; no profile_id → never checked.

2. Dashboard storage meter summed only the first page (≤50) of profiles,
   understating used-bytes + the quota %/bar. The page now walks the
   cursor (fetchAllProfiles, paging at the 100 max while has_more) so the
   meter + bar cover the account's entire profile set.

3. Dashboard cap label mismatch: caps display via fmtBytes, which used
   decimal (1000-based) units while TIER_STORAGE_BYTES_CAP is binary GiB
   (N * 2**30) → "1 GiB" rendered as "1.1 GB" (~7% high). fmtBytes now
   uses binary KiB/MiB/GiB units so the cap shows the true plan number;
   used + cap share the one basis so the % / bar stay correct.

4. Go SDK StorageQuotaExceededError.UsedBytes/CapBytes were int +
   intFromProblem, truncating byte values > 2^31 on 32-bit builds. Now
   int64 via a new int64FromProblem helper.

Full test suite green (0 failed), typecheck clean, Go SDK tests pass.

+++ packages/sdk-go/errors.go
@@ -173,8 +173,12 @@ func (e *QuotaExceededError) Is(target error) bool { return target == ErrQuotaEx
 // and never does.
 type StorageQuotaExceededError struct {
 	apiError
-	UsedBytes int
-	CapBytes  int
+	// UsedBytes/CapBytes are int64: a tier's storage cap is GiB-scale and can
+	// exceed 2^31 bytes (e.g. api_scale = 250 GiB), so a 32-bit `int` would
+	// truncate the value on a 32-bit build (GOARCH=386/arm). int64 is exact on
+	// every target.
+	UsedBytes int64
+	CapBytes  int64
 	Tier      string
 }
